Optimize/cleanup filebufferedchannel a bit.
Make initial zoom size 160 (This is a good balance between worst case scenario were avg size is 1 and smallest zoom is 160, and (the majority) of cases where average size is >=160 (or the smallest zoom level that is <= data_size/2 is >=160).
This best optimizes for the case of mid-size files of a few GB.
We could instead defer making zooms until all or some chromosomes are completely processed (which more exactly aligns to kent tools and has better performance with one thread, but concurrently making zooms gives faster wall times in multi threads), but this has an overhead of either storing the uncompressed data temporarily or uncompressing at the end (since we only do one pass on the input data).
Also, by only doing one pass on the input data, we end up with a lot of "extra" time for smaller zooms anyways. This gap only increases by using libdeflater.